// frontend/packages/project-ide/framework/src/components/resource-folder/type.ts
export type IdType = string;

export enum ResourceTypeEnum {
  Folder = 'folder',
}

export type ItemType = ResourceTypeEnum | string;

export interface ResourceStatusType {
  // Is it a draft?
  draft?: boolean;
  // Error content & number
  problem?: {
    status?: 'normal' | 'error' | 'warning';
    number?: number;
  };
}

export interface ResourceType {
  id: IdType;
  type?: ItemType;
  name: string;
  description?: string;
  children?: ResourceType[];
  path?: string[];
  maxDeep?: number;
  [T: string]: any;
}

export type ResourceMapType<
  T = {
    [T: string]: any;
  },
> = Record<IdType, ResourceType & ResourceStatusType & T>;

export interface CustomResourceConfigType extends ResourceStatusType {
  [T: string]: any;
}

export type CustomResourceConfigMapType = Record<
  IdType,
  CustomResourceConfigType
>;

export type RenderMoreSuffixType =
  | boolean
  | {
      className?: string;
      style?: React.CSSProperties;
      extraProps?: ButtonProps & React.RefAttributes<SemiButton>;
      render?: (
        v: {
          onActive: (e: React.MouseEventHandler<HTMLButtonElement>) => void;
          baseBtn: React.ReactElement;
        } & CommonRenderProps,
      ) => React.ReactElement;
      tooltip?: string | TooltipProps;
    };
// ...
export interface CommonRenderProps {
  resource: ResourceType;
  isSelected?: boolean;
  isTempSelected?: boolean;
  isExpand?: boolean /** This field is only available when resource.type === folder */;
}